+++ src/dotdirdeps.cpp
@@ -383,7 +383,7 @@ void writeDotDirDepGraph(TextStream &t,const DirDef *dd,bool linkRelations)
         size_t nrefs = udir->filePairs().size();
         t << "  " << dir->getOutputFileBase() << "->"
           << usedDir->getOutputFileBase();
-        t << " [headlabel=\"" << static_cast<uint>(nrefs) << "\", labeldistance=1.5";
+        t << " [headlabel=\"" << nrefs << "\", labeldistance=1.5";
         if (linkRelations)
         {
           t << " headhref=\"" << addHtmlExtensionIfMissing(relationName) << "\"";
